requirements.txt
opencv-python-headless
face-recognition
numpy
pandas
pillow
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
